+++ resources/views/components/invitations/create-form.blade.php
@@ -22,9 +22,11 @@
         <p class="mb-4 text-[#F53003] dark:text-[#FF4433]">{{ $message }}</p>
     @enderror
 
-    @error('expires_at')
-        <p class="mb-4 text-[#F53003] dark:text-[#FF4433]">{{ $message }}</p>
-    @enderror
+    @if (config('easy-auth.custom_invitation_expiration'))
+        @error('expires_at')
+            <p class="mb-4 text-[#F53003] dark:text-[#FF4433]">{{ $message }}</p>
+        @enderror
+    @endif
 
     @if (config('easy-auth.multi_use_invitations'))
         @error('max_uses')
@@ -47,15 +49,17 @@
             class="w-full mb-4 px-2 py-1.5 border border-[#19140035] dark:border-[#3E3E3A] rounded-sm text-sm"
         >
 
-        <label for="expires_at" class="block mb-1 text-sm">{{ __('easy-auth::invitations.expires_at') }}</label>
+        @if (config('easy-auth.custom_invitation_expiration'))
+            <label for="expires_at" class="block mb-1 text-sm">{{ __('easy-auth::invitations.expires_at') }}</label>
 
-        <input
-            id="expires_at"
-            name="expires_at"
-            type="datetime-local"
-            value="{{ old('expires_at', $defaultExpiresAt) }}"
-            class="w-full mb-4 px-2 py-1.5 border border-[#19140035] dark:border-[#3E3E3A] rounded-sm text-sm"
-        >
+            <input
+                id="expires_at"
+                name="expires_at"
+                type="datetime-local"
+                value="{{ old('expires_at', $defaultExpiresAt) }}"
+                class="w-full mb-4 px-2 py-1.5 border border-[#19140035] dark:border-[#3E3E3A] rounded-sm text-sm"
+            >
+        @endif
 
         @if (config('easy-auth.multi_use_invitations'))
             <label for="max_uses" class="block mb-1 text-sm">{{ __('easy-auth::invitations.max_uses') }}</label>